requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==1.0.2
gevent==1.3.4
greenlet==0.4.13
gunicorn==19.8.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.1.1
pyasn1==0.4.3
pyasn1-modules==0.2.2
pytz==2018.4
requests==2.19.1
rsa==3.4.2
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
Cython==0.29.17
websocket-client-py3==0.15.0
pydub==0.23.1
